Refactored the existing compiler passes, which detect implementors of certain interfaces and hook up up services to one another, into a generalized version that takes inspiration from OSGi.
Bundles, before passing the compiler pass to the DI container, set it up with information about which of their services "cater for" or "consume" certain interfaces. The compiler pass then checks service definitions for implementations of these interfaces and hooks up the services as either a client or a provider to the bundle's own service.

For instance, a service that wishes to use the Database Service just needs to implement DatabaseServiceAware, and the database bundle will notice that and hook the client up. This relieves client bundles of the responsibility to set up their service definitions correctly. In particular, they don't need to know the service id of the provider anymore.

Thanks to Symfony's caches, this process should bear no performance penalty.
(cherry picked from commit bc2199f)